Move GUI related code into ui subfolder
